remove duplicated fields on Beefy/BeefyItem
